backend/requirements.txt
Flask==2.3.2
requests==2.31.0
openai==0.27.6
beautifulsoup4==4.12.2
python-whois==0.7.3
python-dotenv==1.0.0
